backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
redis==5.1.1
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2